requirements.txt
###############################################
# Core runtime (Python 3.12, CUDA 12.1 target)
###############################################
torch==2.2.0
torchvision==0.17.0

# Numerical / Imaging
numpy>=1.24,<2
pillow>=10,<12
opencv-python-headless>=4.8,<5

# PDF / Export
pypdfium2>=4.30,<5
reportlab>=4.1,<5

###############################################
# Bullseye adapters (local): postproc/export deps
###############################################
pyclipper>=1.3
shapely>=2.0
lxml>=4.9
onnx>=1.15
onnxruntime>=1.17

###############################################
# API / Server (pin to a stable FastAPI set)
###############################################
fastapi==0.111.0
starlette==0.37.2
uvicorn==0.30.6
h11>=0.14,<0.17
click>=8.1,<9
python-multipart>=0.0.9

# Pydantic v2 (compatible pair)
pydantic==2.8.2
pydantic-core==2.20.1
typing-inspection>=0.4
typing-extensions>=4.12

# HTTP utils
requests>=2.28,<3

###############################################
# ML / HF stack
###############################################
huggingface_hub>=0.19,<1
transformers>=4.36,<5
timm==1.0.20
einops==0.8.1
omegaconf==2.3.0
datasets>=2.16,<3
tqdm>=4.65,<5
tensorboard>=2.15,<3
